code for loading sounds done
sounds still need to be triggered (at the moment only the penguin spawn sound is triggered in the GameInputState class). Also for the moment all sounds are a simple click. Appropriate sounds need to be loaded in the assets/Sounds folder and correct paths to these files need to be set in the SoundPlayer class