added new sounds and appropriate tests, also changed some of the ball properties
ball now has higher gravity when going down to speed up the process. maxHeightDiff method of the curve is updated to work for the new range of control points (from 0 to 1 instead of 480 to 0) and new sounds added (more specific to the events e.g. spawn sounds different per obstacle and target and voice sounds added